dbnet-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pillow==11.0.0
numpy==1.26.4
paddlepaddle==3.0.0
paddleocr==2.8.1
opencv-python==4.10.0.84
shapely==2.0.6
